% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Coverage Depth Function for bin}
\usage{
bin(ddn, line, len = 5000)
}
\arguments{
\item{ddn}{a list that contains information about each strain.}

\item{line}{the numeric position in the list for the line of interest.}

\item{len}{the bin size in bp.}
}
\value{
what it outputs
}
\description{
what it does
}
\examples{
#ERR1938053 <- bin(cov.ddn, 11)

}
