% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatScratch.R
\name{formatScratch}
\alias{formatScratch}
\title{Re-formats the results output from ImageJ to sum the multiple measurements from one image into one row.}
\usage{
formatScratch(results)
}
\arguments{
\item{results}{A dataframe that stem from using the Wound_Healing_Tool in ImageJ for the Scratch Assay analysis.}
}
\value{
A .csv file containing the file name broken down into Strain, Replicate, Day and Image, as well as the multiple measurements from one image summed into one row.
}
\description{
Uses one supplied dataframe (scratch_results) from the Scratch Assay ImageJ analysis, and reorganizes the dataframe to include label breakdown names, and sum the multiple measurements into one row.
}
\examples{
# formatScratch(read.csv("Scratch_Assay_Measurements.csv"))

}
