% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatThickness.R
\name{formatThickness}
\alias{formatThickness}
\title{Re-formats the results output from ImageJ to display measurements by Red, Green, Blue and total area.}
\usage{
formatThickness(results)
}
\arguments{
\item{results}{A dataframe that stems from using the Epithelial_Barrier_Thickness_Tool in ImageJ for the Thickness Assay analysis.}
}
\value{
A .csv file containing the file name broken down into Strain, Replicate and Image, as well as the measurements for the colours and total area.
}
\description{
Uses one supplied dataframe (thickness_results) from the Thickness Assay ImageJ analysis, and reorganizes the dataframe to display measurements by the separate colours (Red, Green, Blue) and total area.
}
\examples{
#formatThickness(read.csv("Thickness_Assay_Measurements.csv"))

}
