% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWline.R
\name{SWline}
\alias{SWline}
\title{Coverage Depth Function for SWline}
\usage{
SWline(line, strain_name = NULL)
}
\arguments{
\item{line}{the line of interest, output from bin function}

\item{strain_name}{used for plotting defaults to the name of the line}
}
\value{
what it outputs
}
\description{
What it does
}
