% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthRates.R
\name{growthRates}
\alias{growthRates}
\title{Calculated growth rates for given data}
\usage{
growthRates(d, wells, plot = TRUE, aov = TRUE)
}
\arguments{
\item{d}{A dataframe of the OD values including a column t for time as the first column.}

\item{wells}{A dataframe of the well information including strain type and optionally environment.}

\item{plot}{A logical value indicating whether you would also like to plot your values.}

\item{aov}{A logical value indicated whether you would like to print the anova tests associated.}
}
\value{
A dataframe containing well, lgr, and maxOD information. Additionally, two plots can be created when plot is set to true.
}
\description{
Calculated growth rates for given data
}
