% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthRates.R
\name{growthRates}
\alias{growthRates}
\title{Calculated growth rates for given data}
\usage{
growthRates(d, wells, plotRaw = TRUE, strainName = "Name")
}
\arguments{
\item{d}{A dataframe of the OD values including a column t for time as the first column.}

\item{wells}{A dataframe of the well information including strain type and optionally environment.}

\item{plotRaw}{A logical value indicating whether you would like to plot your raw data.}

\item{strainName}{A string value indicating the strain name.}
}
\value{
A dataframe containing well, lgr, and maxOD information. Additionally, plots of the raw data curves can be created.
}
\description{
Uses two supplied dataframes (or one dataframe and one vector) to calculate growth rate information.
}
